fixed handle delete bitrix standard activity

+++ src/Model/Table/TicketsTable.php
@@ -255,6 +255,11 @@ class TicketsTable extends Table
     public function deleteTicketByActionId($activityId, $memberId)
     {
         $ticket = $this->getByActivityIdAndMemberId($activityId, $memberId);
-        return $this->delete($ticket);
+        if($ticket)
+        {
+            return $this->delete($ticket);
+        } else {
+            return __('Ticket not exist');
+        }
     }
 }
